requirements.txt
streamlit==1.36.0
python-dotenv==1.0.0
openai==1.28.0
anthropic==0.28.0
PyPDF2==4.1.1
python-docx==1.0.1
requests==2.31.0
